Updated generation counter and hash counter to be 64 bit unsigned as they overflowed with k>= 8

+++ src/push_constants.rs
@@ -3,7 +3,7 @@ use vulkano::{DeviceAddress, buffer::BufferContents};
 #[derive(BufferContents, Copy, Clone)]
 #[repr(C)]
 pub struct PushConstants {
-    pub(crate) generation: u32,
+    pub(crate) generation: u64,
     pub(crate) word_count: u32,
     pub(crate) nonce_index: u32,
     pub(crate) k: u32,
